+++ src/control_plane/migration/sql_server/mod.rs
@@ -1,0 +1,5 @@
+mod backup_sql_server_database;
+mod sql_server_backup_options;
+
+pub(crate) use backup_sql_server_database::backup_sql_server_database;
+pub(crate) use sql_server_backup_options::SqlServerBackupOptions;